% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{filter}
\alias{filter}
\title{Filter}
\usage{
filter(x, nsamples, nreads)
}
\arguments{
\item{x}{a matrix.}

\item{nsamples}{number of samples.}

\item{nreads}{number of reads.}
}
\value{
A logical vector indicating whether the gene has to be retained.
}
\description{
Function to filter lowly expressed genes
}
\details{
A gene is retained if it has at least nreads reads in at least nsamples
samples.
}
